+++ test/integration/tests/1659-skip-component/Main.hs
@@ -2,5 +2,5 @@ import StackTest
 
 main :: IO ()
 main = do
-  stack ["build", "--test", "--bench", "--skip-component", "failing-test", "--skip-component", "failing-bench"]
-  stack ["build", ":failing-test", ":failing-bench", "--skip-component", "failing-test", "--skip-component", "failing-bench"]
+  stack ["build", "--test", "--bench", "--skip", "failing-test", "--skip", "failing-bench"]
+  stack ["build", ":failing-test", ":failing-bench", "--skip", "failing-test", "--skip", "failing-bench"]